Move require_environment_variables to flocker.testtools
