+++ flocker/node/agents/testtools.py
@@ -4,19 +4,18 @@
 Test helpers for ``flocker.node.agents``.
 """
 from functools import wraps
-import os
 from uuid import uuid4
 from subprocess import STDOUT, PIPE, Popen
 
 from zope.interface.verify import verifyObject
 from zope.interface import implementer, Interface
 
-from twisted.python.reflect import fullyQualifiedName
 from twisted.python.reflect import prefixedMethodNames
 from twisted.python.filepath import FilePath
-from twisted.trial.unittest import SynchronousTestCase, SkipTest
+from twisted.trial.unittest import SynchronousTestCase
 from twisted.python.components import proxyForInterface
 
+from ..testtools import require_environment_variables
 from .cinder import authenticated_cinder_client
 from .blockdevice import (
     IBlockDeviceAPI, AlreadyAttachedVolume, BlockDeviceVolume, UnknownVolume,
@@ -475,43 +474,6 @@ def make_iblockdeviceapi_tests(blockdevice_api_factory):
             self.api = blockdevice_api_factory(test_case=self)
 
     return Tests
-
-
-def require_environment_variables(required_keys):
-    """
-    Get the values for each of ``keys`` from ``os.environ``.
-
-    :param list keys: The key names to search for in ``os.environ``.
-    :return: a ``dict`` of ``keys`` and corresponding values.
-    :raises: ``SkipTest`` if any of the keys are not found in ``os.environ``.
-    """
-    def decorator(original):
-        missing_keys = []
-        keyvalues = {}
-
-        for key in required_keys:
-            value = os.environ.get(key, None)
-            if value is None:
-                missing_keys.append(key)
-            else:
-                keyvalues[key] = value
-
-        @wraps(original)
-        def wrapper(*args, **kwargs):
-            if missing_keys:
-                raise SkipTest(
-                    '{!r} requires environment variables. '
-                    'Required: "{}". '
-                    'Missing: "{}".'.format(
-                        fullyQualifiedName(original),
-                        '", "'.join(required_keys),
-                        '", "'.join(missing_keys),
-                    )
-                )
-            updated_kwargs = dict(kwargs.items() + keyvalues.items())
-            return original(*args, **updated_kwargs)
-        return wrapper
-    return decorator
 
 
 def require_cinder_credentials(original):
